combine post servlets into one and create marker files for map
