+++ portfolio/src/main/webapp/index.html
@@ -401,7 +401,7 @@
               <img class="icons" src="images/svg/icons8-edit.svg">           
               <h2>contact me</h2>
             </div>
-            <form action="/new-post" method="POST">
+            <form action="/post" method="POST">
               <div class="form-row">
                 <div class="col-md-4">
                   <label for="validationDefault01">first name</label>
